fix for some code review
